refactor category e2e test
