+++ tests/_support/AcceptanceTester.php
@@ -1,7 +1,5 @@
 <?php
 
-use Page\Acceptance\CategoriesPage;
-
 /**
  * Inherited Methods.
  *
@@ -39,20 +37,6 @@ class AcceptanceTester extends Codeception\Actor
         $I->waitForElement('h1.page-header-title', self::TIMEOUT);
         // saving snapshot
         $I->saveSessionSnapshot('login');
-    }
-
-    public function createACategory(string $name): void
-    {
-        $this->amOnPage(CategoriesPage::$URL);
-        $this->waitForElementClickable(CategoriesPage::$NEW_BUTTON);
-        $this->click(CategoriesPage::$NEW_BUTTON);
-        $this->waitForElementClickable(CategoriesPage::$BUNDLE_DROPDOWN);
-        $this->click(CategoriesPage::$BUNDLE_DROPDOWN);
-        $this->waitForElementClickable(CategoriesPage::$BUNDLE_EMAIL_OPTION);
-        $this->click(CategoriesPage::$BUNDLE_EMAIL_OPTION);
-        $this->fillField(CategoriesPage::$TITLE_FIELD, $name);
-        $this->waitForElementClickable(CategoriesPage::$SAVE_AND_CLOSE);
-        $this->click(CategoriesPage::$SAVE_AND_CLOSE);
     }
 
     /**
